toolkit/content/tests/widgets/test_videocontrols_scrubber_position_nopreload.html

<!DOCTYPE HTML>
<!-- Any copyright is dedicated to the Public Domain.
   - https://creativecommons.org/publicdomain/zero/1.0/ -->
<html>
<head>
  <title>Video controls test - Initial scrubber position when preload is turned off</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="head.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>

<div id="content">
  <video width="320" height="240" id="video" mozNoDynamicControls controls="true" preload="none" src="seek_with_sound.ogg"></video>
</div>

<div id="host"></div>

<pre id="test">
<script class="testbody" type="text/javascript">

const video = document.getElementById("video");

add_task(function test_initial_scrubber_position() {
  // Check initial state upon load
  is(video.paused, true, "checking video play state");

  const scrubber = getElementWithinVideo(video, "scrubber");
  ok(scrubber.max, "The max value should be set on the scrubber");
  is(parseInt(scrubber.value), 0, "The initial position should be 0");
});

add_task(async function test_scrubber_after_manual_move() {
  // Kick-start the video before trying to change the scrubber.
  let loadedPromise = new Promise(r => {
    video.addEventListener("canplaythrough", r, {once: true});
  });
  video.play();
  await loadedPromise;
  video.pause();
  const scrubber = getElementWithinVideo(video, "scrubber");
  // Click the middle of the scrubber:
  synthesizeMouseAtCenter(scrubber, {});
  // Expect that the progress updates, too:

  let progress = getElementWithinVideo(video, "progressBar");
  is(
    // toFixed(2) takes care of rounding issues here.
    (progress.value / progress.max).toFixed(2),
    (scrubber.value / scrubber.max).toFixed(2),
    "Should have updated progress bar."
  );
});

</script>
</pre>
</body>
</html>
